requirements.txt
requests
beautifulsoup4
lxml
pandas
boto3
openpyxl
python-dotenv
flask
json